# home-oauth-server/src/main/java/com/softserveinc/ita/homeproject/homeoauthserver/data/AccessToken.java
package com.softserveinc.ita.homeproject.homeoauthserver.data;

import lombok.AllArgsConstructor;
import lombok.Data;

@Data
@AllArgsConstructor
public class AccessToken {

    private String accessToken;
}